chore: move generic errors to application's core
