+++ src/core/errors/not-allowed-error.ts
