+++ src/core/errors/resource-not-found-error.ts
