+++ src/domain/carrier/application/use-cases/delete-courier.ts
@@ -1,6 +1,6 @@
 import { Either, left, right } from '@/core/either'
+import { ResourceNotFoundError } from '@/core/errors/resource-not-found-error'
 import { CouriersRepository } from '../repositories/couriers-repository'
-import { ResourceNotFoundError } from './errors/resource-not-found-error'
 
 interface DeleteCourierUseCaseRequest {
   courierId: string
