+++ src/domain/carrier/application/use-cases/delete-order.ts
@@ -1,6 +1,6 @@
 import { Either, left, right } from '@/core/either'
+import { ResourceNotFoundError } from '@/core/errors/resource-not-found-error'
 import { InMemoryOrdersRepository } from '@test/repositories/in-memory/in-memory-orders-repository'
-import { ResourceNotFoundError } from './errors/resource-not-found-error'
 
 interface DeleteOrderUseCaseRequest {
   orderId: string
