+++ src/domain/carrier/application/use-cases/get-order-by-id.ts
@@ -1,7 +1,7 @@
 import { Either, left, right } from '@/core/either'
+import { ResourceNotFoundError } from '@/core/errors/resource-not-found-error'
 import { InMemoryOrdersRepository } from '@test/repositories/in-memory/in-memory-orders-repository'
 import { Order } from '../../enterprise/entities/order'
-import { ResourceNotFoundError } from './errors/resource-not-found-error'
 
 interface GetOrderByIdUseCaseRequest {
   orderId: string
