+++ src/domain/carrier/application/use-cases/update-order-courier.spec.ts
@@ -1,9 +1,9 @@
 import { UniqueEntityID } from '@/core/entities/unique-entity-id'
+import { ResourceNotFoundError } from '@/core/errors/resource-not-found-error'
 import { makeCourier } from '@test/factories/make-courier'
 import { makeOrder } from '@test/factories/make-order'
 import { InMemoryCouriersRepository } from '@test/repositories/in-memory/in-memory-couriers-repository'
 import { InMemoryOrdersRepository } from '@test/repositories/in-memory/in-memory-orders-repository'
-import { ResourceNotFoundError } from './errors/resource-not-found-error'
 import { UpdateOrderCourierUseCase } from './update-order-courier'
 
 let inMemoryOrdersRepository: InMemoryOrdersRepository
