+++ test/repositories/in-memory/in-memory-notifications-repository.ts
@@ -5,9 +5,9 @@ export class InMemoryNotificationsRepository
   implements NotificationsRepository
 {
   items: Notification[] = []
-  
+
   async findById(id: string): Promise<Notification | null> {
-    const notification = this.items.find(item => item.id.toValue() === id)
+    const notification = this.items.find((item) => item.id.toValue() === id)
 
     if (!notification) {
       return null
@@ -15,13 +15,15 @@ export class InMemoryNotificationsRepository
 
     return notification
   }
-  
+
   async create(notification: Notification): Promise<void> {
     this.items.push(notification)
   }
 
   async save(notification: Notification): Promise<void> {
-    const notificationIndex = this.items.findIndex((item) => item.id === notification.id)
+    const notificationIndex = this.items.findIndex(
+      (item) => item.id === notification.id,
+    )
 
     this.items[notificationIndex] = notification
   }
